fixes to pull request conversations

+++ src/Model/GameObject/NPC.cpp
@@ -99,7 +99,7 @@ void NPC::save_object() {
 	ObjectSaving::addValue("moveSpeed", m_move_speed, true);
 	ObjectSaving::closeStruct();
 
-	for (int count = 0; count < pb->colliderSize(); count++) {
+	for (size_t count = 0; count < pb->colliderSize(); count++) {
 		int type = pb->getColliderType(count);
 		ObjectSaving::createStruct("collider" + std::to_string(count + 1));
 		saveCollider(count, type);
